feat(calculator): array and string based implemtation with the same tests
