+++ src/calculator/string-based.spec.js
@@ -1,0 +1,13 @@
+'use strict';
+
+const Calculator = require('./string-based');
+const tests = require('./shared-tests');
+
+describe('String based Calculator', function() {
+
+  beforeEach(function() {
+    this.subject = Calculator.create();
+  });
+
+  tests();
+});
